Move pipeline logs command
